Change MonoType.d.ts to be MonoTypes.ts and update all references
