+++ src/Components/Web.JS/src/Boot.WebAssembly.ts
@@ -9,6 +9,7 @@ import { setEventDispatcher } from './Rendering/RendererEventDispatcher';
 import { WebAssemblyResourceLoader } from './Platform/WebAssemblyResourceLoader';
 import { WebAssemblyConfigLoader } from './Platform/WebAssemblyConfigLoader';
 import { BootConfigResult } from './Platform/BootConfig';
+import { Pointer } from './Platform/Platform';
 
 let started = false;
