+++ src/Components/Web.JS/src/Platform/Platform.ts
@@ -19,3 +19,12 @@ export interface Platform {
   readStringField(baseAddress: Pointer, fieldOffset?: number): string | null;
   readStructField<T extends Pointer>(baseAddress: Pointer, fieldOffset?: number): T;
 }
+
+// We don't actually instantiate any of these at runtime. For perf it's preferable to
+// use the original 'number' instances without any boxing. The definitions are just
+// for compile-time checking, since TypeScript doesn't support nominal types.
+export interface MethodHandle { MethodHandle__DO_NOT_IMPLEMENT: any }
+export interface System_Object { System_Object__DO_NOT_IMPLEMENT: any }
+export interface System_String extends System_Object { System_String__DO_NOT_IMPLEMENT: any }
+export interface System_Array<T> extends System_Object { System_Array__DO_NOT_IMPLEMENT: any }
+export interface Pointer { Pointer__DO_NOT_IMPLEMENT: any }
