+++ src/Components/Web.JS/src/Platform/WebAssemblyConfigLoader.ts
@@ -1,4 +1,6 @@
 import { BootConfigResult } from './BootConfig';
+import { BINDING } from './Mono/MonoTypes';
+import { System_String, Pointer } from './Platform';
 
 export class WebAssemblyConfigLoader {
   static async initAsync(bootConfigResult: BootConfigResult): Promise<void> {
